Struct tor_circmgr::hspool::HsCircPool

pub struct HsCircPool<R: Runtime> { /* private fields */ }

An object to provide circuits for implementing onion services.

Implementations

impl<R: Runtime> HsCircPool<R>

pub fn new(circmgr: &Arc<CircMgr<R>>) -> Arc<Self>

Create a new HsCircPool.

This will not work properly before “launch_background_tasks” is called.

pub fn launch_background_tasks( self: &Arc<Self>, runtime: &R, netdir_provider: &Arc<dyn NetDirProvider + 'static> ) -> Result<Vec<TaskHandle>>

Launch the periodic daemon tasks required by the manager to function properly.

Returns a set of TaskHandles that can be used to manage the daemon tasks.

pub async fn get_or_launch_client_rend<'a>( &self, netdir: &'a NetDir ) -> Result<(Arc<ClientCirc>, Relay<'a>)>

Create a circuit suitable for use as a rendezvous circuit by a client.

Return the circuit, along with a Relay from netdir representing its final hop.

Only makes a single attempt; the caller needs to loop if they want to retry.

pub async fn get_or_launch_specific<T>( &self, netdir: &NetDir, kind: HsCircKind, target: T ) -> Result<Arc<ClientCirc>>where T: CircTarget,

Create a circuit suitable for use for kind, ending at the chosen hop target.

Only makes a single attempt; the caller needs to loop if they want to retry.

pub fn estimate_timeout(&self, timeout_action: &Action) -> Duration

Return an estimate-based delay for how long a given Action should be allowed to complete.

This function has the same semantics as CircMgr::estimate_timeout. See the notes there.

In particular you do not need to use this function in order to get reasonable timeouts for the circuit-building operations provided by HsCircPool.